+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [v0.1.0]
+### Added
+- A method `getCppWLC2D` was added to helpers.py for generating
+  two-dimensional wormlike chains.
+		
 ## [v0.0.1]
 ### Fixed
 - The Python interface to the self-avoiding wormlike chain no longer
@@ -20,4 +25,5 @@ All notable changes to this project will be documented in this file.
 - Fixed off-by-one error in the wormlike chain generation code.
 - Errors in computation of chain statistics
 
+[v0.1.0]: https://github.com/kmdouglass/PolymerCpp/compare/v0.0.1...v0.1.0
 [v0.0.1]: https://github.com/kmdouglass/PolymerCpp/compare/v0.0.0...v0.0.1